+++ shield-update.sh
@@ -26,7 +26,6 @@ CURRENT_DIR=$(cd $(dirname $0); pwd)
 cd $CURRENT_DIR
 SCRIPTS_URL="https://ericom-tec.ashisuto.co.jp/shield"
 
-
 if [ -f .es_branch ]; then
     BRANCH=$(cat .es_branch)
 elif [ -f ${ES_PATH}/.es_branch ]; then