+++ sup/shield-sup.sh
@@ -5,7 +5,7 @@
 
 ####################
 ### K.K. Ashisuto
-### VER=20200226a
+### VER=20200313a
 ####################
 
 ####-----------------
@@ -28,9 +28,7 @@ if [ ! -e $ES_PATH ];then
     mkdir -p $ES_PATH/sup
 fi
 
-
 SCRIPTS_URL="https://ericom-tec.ashisuto.co.jp/shield"
-
 
 usage() {
    echo "$0 [-y]"
@@ -226,6 +224,14 @@ if [ -f shield-start.sh ];then
     cp ./.es*  $TMPDIR/shield/ 2>/dev/null
     cp ./.ra*  $TMPDIR/shield/ 2>/dev/null
     cp ./*.sh  $TMPDIR/shield/ 2>/dev/null
+
+    #/var/lib/docker/containers log
+    mkdir -p $TMPDIR/varlibdokcer-logs
+    for TARGET in `ls /var/lib/docker/containers`
+    do
+            NAME=$(cat /var/lib/docker/containers/${TARGET}/config.v2.json | jq .Name | sed -e s/\"//g | sed -e s"/\///")
+            cp /var/lib/docker/containers/${TARGET}/${TARGET}-json.log $TMPDIR/varlibdokcer-logs/${NAME}-json.log
+    done
 fi
 
 # for swarm